refactor(1013): extract TelemetryEmitter to src/analytics/ (SC-001, position 9)
- Move TelemetryEmitter (~215 LOC) from MistHelper.py:17773-17987 to
  src/analytics/telemetry_emitter.py; class is stdlib-only + progress_event
  dataclasses (no MistHelper module-global dependencies, so no lazy
  importlib.import_module pattern needed).
- Hoist import glob to module top (was previously inside enforce_retention).
- Add re-export in MistHelper.py imports block so 4 caller sites at
  18416/18419/18498/18849 resolve transparently.
- Drop now-unused TestSummary import from MistHelper.py; ProgressContext
  still consumed by ~10 emit_progress_complete callers.
- Add breadcrumb comment at prior class location.

Local gate: black --check + ruff clean; py_compile OK; --test 66/66 pass.
